requirements.txt
black
dash
flask
jupyter
matplotlib
mypy
numpy
pandas
plotly
pycodestyle
python-dateutil
ruff 
sqlalchemy
tqdm
polars
connectorx
pyarrow